requirements.txt
aiomysql==0.1.1
anyio==3.6.2
APScheduler==3.9.1.post1
async-sqlalchemy==1.0.0
asyncpg==0.27.0
black==22.12.0
certifi==2022.12.7
click==8.1.3
emoji==2.2.0
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.2
PyMySQL==1.0.2
python-dotenv==0.21.0
python-telegram-bot==20.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
